server/requirements.txt
falcon==1.0.0
falcon-cors==1.0.1
google-apputils==0.4.2
gunicorn==19.6.0
protobuf==3.0.0b3
python-dateutil==2.5.3
python-gflags==3.0.5
python-mimeparse==1.5.2
pytz==2016.6.1
six==1.10.0
